+++ src/main/scala/io/bartholomews/fsclient/codecs/CodecSyntax.scala
@@ -27,6 +27,10 @@ trait CodecSyntax extends PlainTextDecodingSyntax {
 
   implicit def urlFormEntityEncoder[F[_]]: EntityEncoder[F, UrlForm] = UrlForm.entityEncoder
 
+  implicit val rawUnitPipe: RawDecoder[Unit] = new RawDecoder[Unit] {
+    override def decode[F[_]: Effect]: Pipe[F, Response[F], Unit] = _.map(_ => ())
+  }
+
   implicit val rawPlainTextPipe: RawDecoder[String] = new RawDecoder[String] {
     override def decode[F[_]: Effect]: Pipe[F, Response[F], String] =
       _.flatMap { response =>
@@ -47,6 +51,8 @@ trait CodecSyntax extends PlainTextDecodingSyntax {
 
   // implicit def decodeIdentity[F[_], A]: Pipe[F, A, A] = _.map(identity)
   implicit def stringDecoderPipe[F[_]]: Pipe[F, String, String] = _.map(identity)
+
+  implicit def decodeAsUnit[F[_], A]: Pipe[F, A, Unit] = _.map(_ => ())
 
   implicit def decodeJsonAsString[F[_]: Sync]: Pipe[F, Json, String] =
     deriveJsonPipe[F, String](implicitly[Sync[F]], Decoder.decodeString)