+++ src/main/scala/io/bartholomews/fsclient/requests/FsAuthRequest.scala
@@ -7,7 +7,7 @@ import io.bartholomews.fsclient.codecs.RawDecoder
 import io.bartholomews.fsclient.entities.oauth.Signer
 import io.bartholomews.fsclient.utils.HttpTypes.HttpResponse
 import io.circe.Json
-import org.http4s.Method.{DefaultMethodWithBody, SafeMethodWithBody}
+import org.http4s.Method.{DefaultMethodWithBody, IdempotentMethodWithBody, SafeMethodWithBody}
 import org.http4s.{EntityEncoder, Method}
 
 sealed trait FsAuthRequest[Body, Raw, Res] extends FsClientRequest[Body] {
@@ -22,8 +22,15 @@ sealed trait FsAuthRequest[Body, Raw, Res] extends FsClientRequest[Body] {
 
 object FsAuthRequest {
 
+  type SimplePost[Body] = Post[Body, Unit, Unit]
+  type SimplePut[Body] = Put[Body, Unit, Unit]
+
   trait Get[Body, Raw, Res] extends FsAuthRequest[Body, Raw, Res] {
     final override private[fsclient] def method: SafeMethodWithBody = Method.GET
+  }
+
+  trait Put[Body, Raw, Res] extends FsAuthRequest[Body, Raw, Res] {
+    final override private[fsclient] def method: IdempotentMethodWithBody = Method.PUT
   }
 
   trait Post[Body, Raw, Res] extends FsAuthRequest[Body, Raw, Res] {
@@ -36,6 +43,11 @@ object AuthJsonRequest {
     final override private[fsclient] def body: Option[Nothing] = None
   }
 
+  trait Put[Body, Res] extends FsAuthRequest.Put[Body, Json, Res] {
+    def entityBody: Body
+    final override private[fsclient] def body: Option[Body] = Some(entityBody)
+  }
+
   trait Post[Body, Res] extends FsAuthRequest.Post[Body, Json, Res] {
     def entityBody: Body
     final override private[fsclient] def body: Option[Body] = Some(entityBody)
@@ -46,6 +58,12 @@ object AuthPlainTextRequest {
   trait Get[Res] extends FsAuthRequest.Get[Nothing, String, Res] {
     final override private[fsclient] def body: Option[Nothing] = None
   }
+
+  trait Put[Body, Res] extends FsAuthRequest.Put[Body, String, Res] {
+    def entityBody: Body
+    final override private[fsclient] def body: Option[Body] = Some(entityBody)
+  }
+
   trait Post[Body, Res] extends FsAuthRequest.Post[Body, String, Res] {
     def entityBody: Body
     final override private[fsclient] def body: Option[Body] = Some(entityBody)