+++ CHANGELOG.md
@@ -8,10 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `FsJsonRequest[A]` trait which imports an `EntityEncoder`
+- `PUT` requests and `Unit` decoders
 ### Removed
 - Logger name from config (to avoid concurrency error when using multiple clients)
 ### Changed
 - Updated [Readme](https://github.com/bartholomews/fsclient/compare/v0.0.2...HEAD#diff-04c6e90faac2675aa89e2176d2eec7d8) scala snippets
+- Decode < 400 as `success` status instead of just 200
 ### Fixed
 
 ## [0.0.2] - 2020-05-23